Moved article backup to useForm hook, refactored BackupNotice component
