+++ resources/js/Hooks/Article/useForm.js
@@ -1,4 +1,6 @@
+import {onMounted, ref} from "vue";
 import {useForm as useInertiaForm} from "@inertiajs/vue3";
+import moment from "moment";
 
 function useForm(article = null) {
   const form = useInertiaForm({
@@ -13,7 +15,28 @@ function useForm(article = null) {
     status: null,
   });
 
-  const submit = (params) => {
+  const backup = ref(null);
+
+  const update = (prop, value) => {
+    form[prop] = value;
+
+    localStorage.setItem('article_backup', JSON.stringify({
+      ...form.data(),
+      save_date: moment().format(),
+    }));
+  };
+
+  const restore = () => {
+    form.text = backup.value.text;
+    form.title = backup.value.title;
+    form.summary = backup.value.summary;
+    form.tags = backup.value.tags;
+    form.topics = backup.value.topics;
+    form.media = backup.value.media;
+    form.difficulty = backup.value.difficulty;
+  };
+
+  const send = (params) => {
     const uri = article
       ? route('article.update', {article: article.id})
       : route('article.create');
@@ -28,9 +51,20 @@ function useForm(article = null) {
       .post(uri, params);
   };
 
+  onMounted(() => {
+    const data = localStorage.getItem('article_backup');
+
+    if (data) {
+      backup.value = JSON.parse(data);
+    }
+  });
+
   return {
     form,
-    submit,
+    backup,
+    update,
+    restore,
+    send,
   };
 }
